port two piping tests to WPT
This includes the last user of several helper functions
which had already been ported over to WPT in
commit fd44398d410beb6879fb27290e690603bed04c4d.

Though sequentialReadableStream was ported to WPT in that commit,
it was never actually USED until now.
