--- reference-implementation/test/utils/duck-typed-pass-through-transform.js
@@ -1,25 +1,0 @@
-'use strict';
-
-module.exports = () => {
-  let enqueueInReadable;
-  let closeReadable;
-
-  return {
-    writable: new WritableStream({
-      write(chunk) {
-        enqueueInReadable(chunk);
-      },
-
-      close() {
-        closeReadable();
-      }
-    }),
-
-    readable: new ReadableStream({
-      start(c) {
-        enqueueInReadable = c.enqueue.bind(c);
-        closeReadable = c.close.bind(c);
-      }
-    })
-  };
-};
