+++ reference-implementation/to-upstream-wpts/.eslintrc.json
@@ -56,6 +56,7 @@
     "methodRejects": true,
     "readableStreamToArray": true,
     "recordingReadableStream": true,
-    "recordingWritableStream": true
+    "recordingWritableStream": true,
+    "sequentialReadableStream": true
   }
 }
