+++ reference-implementation/to-upstream-wpts/piping/general.https.html
@@ -6,6 +6,7 @@
 <script src="../resources/test-initializer.js"></script>
 <script src="../resources/test-utils.js"></script>
 <script src="../resources/recording-streams.js"></script>
+<script src="../resources/rs-utils.js"></script>
 
 <script src="general.js"></script>
 <script>
